backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aegiscopilot-backend"
version = "0.1.0"
description = "Backend service for the AegisCopilot enterprise knowledge-base AI agent"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "alembic>=1.13.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "psycopg[binary]>=3.2.3",
  "python-multipart>=0.0.9",
  "pypdf>=5.4.0",
  "cryptography>=3.1",
  "python-docx>=1.1.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]
milvus = [
  "pymilvus>=2.5.0",
]
ocr = [
  "Pillow>=10.0.0",
  "pytesseract>=0.3.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
